Autorizacion por roles en products-service

+++ Server/products-service/src/main/java/com/productsservice/controller/ProductController.java
@@ -2,7 +2,11 @@ package com.productsservice.controller;
 
 import com.productsservice.model.Product;
 import com.productsservice.service.IProductService;
+import feign.FeignException;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,29 +20,62 @@ public class ProductController {
     @Autowired
     private IProductService productService;
 
+    // ADMIN
     @PostMapping("")
-    public Product createProduct(@RequestBody Product product){
-        return productService.createProduct(product);
+    public ResponseEntity<?> createProduct(@RequestBody Product product, HttpServletRequest request){
+        if(!hasAdminRole(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+        return new ResponseEntity<>(productService.createProduct(product), HttpStatus.OK);
     }
 
+    // ALL
     @GetMapping("")
-    public List<Product> getProducts(){
-        return productService.getProducts();
+    public ResponseEntity<List<Product>> getProducts(){
+        return new ResponseEntity<>(productService.getProducts(), HttpStatus.OK);
     }
 
+    // ALL
     @GetMapping("/{id_product}")
-    public Product getProductById(@PathVariable Long id_product){
-        return productService.getProductById(id_product);
+    public ResponseEntity<Product> getProductById(@PathVariable Long id_product){
+        return new ResponseEntity<>(productService.getProductById(id_product), HttpStatus.OK);
     }
 
+    // ADMIN
     @PutMapping("")
-    public Product editProduct(@RequestBody Product product){
-        return productService.editProduct(product);
+    public ResponseEntity<?> editProduct(@RequestBody Product product, HttpServletRequest request){
+        if(!hasAdminRole(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+        return new ResponseEntity<>(productService.editProduct(product), HttpStatus.OK);
     }
 
+    // ADMIN
     @DeleteMapping("/{id_product}")
-    public void deleteProductById(@PathVariable Long id_product){
+    public ResponseEntity<?> deleteProductById(@PathVariable Long id_product, HttpServletRequest request){
+        if(!hasAdminRole(request)){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         productService.deleteProductById(id_product);
+        return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    private String extractTokenFromHeader(HttpServletRequest request){
+        String authorizationHeader = request.getHeader("Authorization");
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            return authorizationHeader.substring(7);
+        }
+        return null;
+    }
+
+    private boolean hasAdminRole(HttpServletRequest request){
+        String token = extractTokenFromHeader(request);
+        try{
+            String role = String.valueOf(productService.getRoleByToken(token));
+            return "ADMIN".equals(role);
+        } catch (FeignException.Unauthorized e){
+            return false;
+        }
     }
 
 }
